requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-embed-video==1.4.10
easy-thumbnails==2.10
easyocr==1.7.2
filelock==3.17.0
fsspec==2024.12.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.5
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
paperclip==2.7.2
pillow==11.1.0
pyclipper==1.3.0.post6
pyperclip==1.9.0
pytesseract==0.3.13
python-bidi==0.6.3
python-magic==0.4.27
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.1
scipy==1.15.1
setuptools==75.8.0
shapely==2.0.6
sqlparse==0.5.3
sympy==1.13.1
tifffile==2025.1.10
tk==0.1.0
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
